Title: 구간의 합이 같게 나누기, Time: 1480ms, Memory: 35MB, Status: Wrong Answer - Codetree

+++ 250505/구간의 합이 같게 나누기/divide-by-equal-sum-of-intervals.java
@@ -16,7 +16,7 @@ public class Main {
 
         arr = new long[n + 2];
         for (int i = 1; i <= n; i ++) {
-            arr[i] = sc.nextInt();
+            arr[i] = sc.nextLong();
         }
 
         left = new long[n + 2];
